requirements.txt
selenium==3.141.0
behave==1.2.5
pyhamcrest==1.9.0
requests==2.24.0
python-dotenv==0.9.1
pylint==2.2.2
exchangelib==3.1.1